remove sonar bugs

+++ src/main/java/online/kheops/auth_server/util/Consts.java
@@ -31,7 +31,7 @@ public class Consts {
     }
 
     public static final int NUMBER_OF_RETRY_WEBHOOK = 5;
-    public static final int SECONDE_BEFORE_RETRY_WEBHOOK = 5;
+    public static final long SECONDE_BEFORE_RETRY_WEBHOOK = 5;
     public static final List<String> VALID_SCHEMES_WEBHOOK_URL = List.of("http", "https");
 
     public static final String HOST_ROOT_PARAMETER = "online.kheops.root.uri";
